refactor: split large files exceeding size guidelines (#102)
Split three critical files that significantly exceeded the project's size
guidelines (150 lines for utilities, 200-300 for components):

**lib/connections/plex.ts (1,430 lines → 7 files)**
- plex-core.ts - Core types and utilities
- plex-connection.ts - Connection testing and server identity
- plex-users.ts - User management functions
- plex-invitations.ts - Invitation and sharing functions
- plex-sessions.ts - Session and library management
- plex-search.ts - Search functionality
- plex-marking.ts - Mark watched/unwatched

**lib/wrapped/statistics.ts (1,289 lines → 4 files)**
- statistics-types.ts - Type definitions
- tautulli-statistics.ts - Tautulli API functions
- overseerr-statistics.ts - Overseerr API functions
- server-statistics.ts - Plex server statistics

**actions/admin.ts (1,241 lines → 5 files)**
- admin-config.ts - Config and wrapped settings
- admin-llm.ts - LLM usage and providers
- admin-servers.ts - Server configurations
- admin-users.ts - User management
- admin-settings.ts - Combined settings fetch

All original files now serve as barrel files that re-export from
their domain-specific modules, maintaining backward compatibility.
